keep moving

+++ src/ConnectionCopyingInputStream.cpp
@@ -17,7 +17,7 @@ namespace rprotobuf{
 	int	ConnectionCopyingInputStream::Read(void * buffer, int size){
 		
 		/* TODO: error handling */
-		Rcpp::RawVector res = readBin( con, Rcpp::RawVector( (size_t)0 ), size ) ;
+		Rcpp::RawVector res = readBin( con, Rcpp::RawVector( 0 ), size ) ;
 		memcpy( buffer, res.begin() , res.size() ) ;
 		return res.size() ;
 	}
